Split benchmark and generate_wats in different files
